+++ crates/mirror-mirror-macros/src/derive_reflect/enum_.rs
@@ -9,6 +9,7 @@ use syn::Type;
 
 use super::attrs::InnerAttrs;
 use super::attrs::ItemAttrs;
+use super::trivial_reflect_methods;
 use super::Generics;
 use crate::stringify;
 
@@ -21,6 +22,7 @@ pub(super) fn expand(
     let variants = VariantData::try_from_enum(&enum_)?;
 
     let describe_type = expand_describe_type(ident, &variants, &attrs, generics);
+    let default_value = expand_default_value(ident, &attrs, generics);
     let reflect = expand_reflect(ident, &variants, &attrs, generics)?;
     let from_reflect = (!attrs.from_reflect_opt_out)
         .then(|| expand_from_reflect(ident, &variants, &attrs, generics));
@@ -28,6 +30,7 @@ pub(super) fn expand(
 
     Ok(quote! {
         #describe_type
+        #default_value
         #reflect
         #from_reflect
         #enum_
@@ -102,7 +105,6 @@ fn expand_describe_type(
 
     let meta = attrs.meta();
     let docs = attrs.docs();
-
     let Generics {
         impl_generics,
         type_generics,
@@ -117,6 +119,37 @@ fn expand_describe_type(
                     EnumNode::new::<Self>(variants, #meta, #docs)
                 })
             }
+        }
+    }
+}
+
+fn expand_default_value(ident: &Ident, attrs: &ItemAttrs, generics: &Generics<'_>) -> TokenStream {
+    let Generics {
+        impl_generics,
+        type_generics,
+        where_clause,
+    } = generics;
+
+    let fn_default_value = if attrs.default_opt_out {
+        quote! {
+            fn default_value() -> Option<Value> {
+                None
+            }
+        }
+    } else {
+        quote! {
+            fn default_value() -> Option<Value> {
+                fn __default<Z: Default>() -> Z {
+                    Default::default()
+                }
+                Some(__default::<#ident #type_generics>().to_value())
+            }
+        }
+    };
+
+    quote! {
+        impl #impl_generics DefaultValue for #ident #type_generics #where_clause {
+            #fn_default_value
         }
     }
 }
@@ -138,7 +171,7 @@ fn expand_reflect(
                         let ident = field.ident;
                         let ident_string = stringify(ident);
                         quote! {
-                            if let Some(new_value) = enum_.field(#ident_string) {
+                            if let Some(new_value) = __mirror_mirror_enum.field(#ident_string) {
                                 #ident.patch(new_value);
                             }
                         }
@@ -157,7 +190,7 @@ fn expand_reflect(
                         |(index, field)| {
                             let ident = &field.fake_ident;
                             quote! {
-                                if let Some(new_value) = enum_.field_at(#index) {
+                                if let Some(new_value) = __mirror_mirror_enum.field_at(#index) {
                                     #ident.patch(new_value);
                                 }
                             }
@@ -184,12 +217,12 @@ fn expand_reflect(
 
         if attrs.clone_opt_out {
             quote! {
-                fn patch(&mut self, value: &dyn Reflect) {
-                    if let Some(enum_) = value.reflect_ref().as_enum() {
-                        if let Some(new) = FromReflect::from_reflect(value) {
+                fn patch(&mut self, __mirror_mirror_value: &dyn Reflect) {
+                    if let Some(__mirror_mirror_enum) = __mirror_mirror_value.reflect_ref().as_enum() {
+                        if let Some(new) = FromReflect::from_reflect(__mirror_mirror_value) {
                             *self = new;
                         } else {
-                            let variant_matches = self.variant_name() == enum_.variant_name();
+                            let variant_matches = self.variant_name() == __mirror_mirror_enum.variant_name();
                             match self {
                                 #(#match_arms)*
                                 _ => {}
@@ -200,14 +233,14 @@ fn expand_reflect(
             }
         } else {
             quote! {
-                fn patch(&mut self, value: &dyn Reflect) {
-                    if let Some(new) = value.downcast_ref::<Self>() {
+                fn patch(&mut self, __mirror_mirror_value: &dyn Reflect) {
+                    if let Some(new) = __mirror_mirror_value.downcast_ref::<Self>() {
                         *self = new.clone();
-                    } else if let Some(enum_) = value.reflect_ref().as_enum() {
-                        if let Some(new) = FromReflect::from_reflect(value) {
+                    } else if let Some(__mirror_mirror_enum) = __mirror_mirror_value.reflect_ref().as_enum() {
+                        if let Some(new) = FromReflect::from_reflect(__mirror_mirror_value) {
                             *self = new;
                         } else {
-                            let variant_matches = self.variant_name() == enum_.variant_name();
+                            let variant_matches = self.variant_name() == __mirror_mirror_enum.variant_name();
                             match self {
                                 #(#match_arms)*
                                 _ => {}
@@ -231,7 +264,10 @@ fn expand_reflect(
                         let ident = &field.ident;
                         let ident_string = stringify(ident);
                         quote! {
-                            value.set_struct_field(#ident_string, #ident.to_value());
+                            __mirror_mirror_value.set_struct_field(
+                                #ident_string,
+                                #ident.to_value(),
+                            );
                         }
                     });
 
@@ -239,9 +275,9 @@ fn expand_reflect(
 
                     quote! {
                         Self::#variant_ident { #(#field_names,)* } => {
-                            let mut value = EnumValue::new_struct_variant_with_capacity(#variant_ident_string, #fields_len);
+                            let mut __mirror_mirror_value = EnumValue::new_struct_variant_with_capacity(#variant_ident_string, #fields_len);
                             #(#set_fields)*
-                            value.finish().into()
+                            __mirror_mirror_value.finish().into()
                         }
                     }
                 }
@@ -257,11 +293,11 @@ fn expand_reflect(
 
                     quote! {
                         Self::#variant_ident(#(#field_names,)*) => {
-                            let mut value = EnumValue::new_tuple_variant_with_capacity(#variant_ident_string, #fields_len);
+                            let mut __mirror_mirror_value = EnumValue::new_tuple_variant_with_capacity(#variant_ident_string, #fields_len);
                             #(
-                                value.push_tuple_field(#included_fields.to_value());
+                                __mirror_mirror_value.push_tuple_field(#included_fields.to_value());
                             )*
-                            value.finish().into()
+                            __mirror_mirror_value.finish().into()
                         }
                     }
                 }
@@ -294,12 +330,6 @@ fn expand_reflect(
         }
     };
 
-    let fn_type_info = quote! {
-        fn type_descriptor(&self) -> Cow<'static, TypeDescriptor> {
-            <Self as DescribeType>::type_descriptor()
-        }
-    };
-
     let fn_debug = attrs.fn_debug_tokens();
     let fn_clone_reflect = attrs.fn_clone_reflect_tokens();
 
@@ -309,25 +339,12 @@ fn expand_reflect(
         where_clause,
     } = generics;
 
+    let trivial_reflect_methods = trivial_reflect_methods();
+
     Ok(quote! {
         impl #impl_generics Reflect for #ident #type_generics #where_clause {
-            fn as_any(&self) -> &dyn Any {
-                self
-            }
+            #trivial_reflect_methods
 
-            fn as_any_mut(&mut self) -> &mut dyn Any {
-                self
-            }
-
-            fn as_reflect(&self) -> &dyn Reflect {
-                self
-            }
-
-            fn as_reflect_mut(&mut self) -> &mut dyn Reflect {
-                self
-            }
-
-            #fn_type_info
             #fn_patch
             #fn_to_value
             #fn_clone_reflect
@@ -373,21 +390,21 @@ fn expand_from_reflect(
                         if let Some(from_reflect_with) = field.from_reflect_with() {
                             quote! {
                                 #ident: {
-                                    let value = enum_.field(#ident_string)?;
+                                    let value = __mirror_mirror_enum.field(#ident_string)?;
                                     #from_reflect_with(value)?
                                 },
                             }
                         } else if attrs.clone_opt_out {
                             quote! {
                                 #ident: {
-                                    let value = enum_.field(#ident_string)?;
+                                    let value = __mirror_mirror_enum.field(#ident_string)?;
                                     FromReflect::from_reflect(value)?
                                 },
                             }
                         } else {
                             quote! {
                                 #ident: {
-                                    let value = enum_.field(#ident_string)?;
+                                    let value = __mirror_mirror_enum.field(#ident_string)?;
                                     if let Some(value) = value.downcast_ref::<#ty>() {
                                         value.to_owned()
                                     } else {
@@ -416,21 +433,21 @@ fn expand_from_reflect(
                         if let Some(from_reflect_with) = field.from_reflect_with() {
                             quote! {
                                 {
-                                    let value = enum_.field_at(#idx)?;
+                                    let value = __mirror_mirror_enum.field_at(#idx)?;
                                     #from_reflect_with(value)?
                                 },
                             }
                         } else if attrs.clone_opt_out {
                             quote! {
                                 {
-                                    let value = enum_.field_at(#idx)?;
+                                    let value = __mirror_mirror_enum.field_at(#idx)?;
                                     FromReflect::from_reflect(value)?
                                 },
                             }
                         } else {
                             quote! {
                                 {
-                                    let value = enum_.field_at(#idx)?;
+                                    let value = __mirror_mirror_enum.field_at(#idx)?;
                                     if let Some(value) = value.downcast_ref::<#ty>() {
                                         value.to_owned()
                                     } else {
@@ -466,9 +483,9 @@ fn expand_from_reflect(
 
     quote! {
         impl #impl_generics FromReflect for #ident #type_generics #where_clause {
-            fn from_reflect(reflect: &dyn Reflect) -> Option<Self> {
-                let enum_ = reflect.reflect_ref().as_enum()?;
-                match enum_.variant_name() {
+            fn from_reflect(__mirror_mirror_reflect: &dyn Reflect) -> Option<Self> {
+                let __mirror_mirror_enum = __mirror_mirror_reflect.reflect_ref().as_enum()?;
+                match __mirror_mirror_enum.variant_name() {
                     #(#match_arms)*
                     _ => None,
                 }
@@ -538,7 +555,7 @@ fn expand_enum(
                             let ident = &field.ident;
                             let ident_string = stringify(ident);
                             quote! {
-                                if name == #ident_string {
+                                if __mirror_mirror_name == #ident_string {
                                     return Some(#ident);
                                 }
                             }
@@ -561,7 +578,7 @@ fn expand_enum(
 
         quote! {
             #[allow(unused_variables, unreachable_code)]
-            fn field(&self, name: &str) -> Option<&dyn Reflect> {
+            fn field(&self, __mirror_mirror_name: &str) -> Option<&dyn Reflect> {
                 match self {
                     #(#match_arms)*
                     _ => {}
@@ -586,7 +603,7 @@ fn expand_enum(
                             let ident = &field.ident;
                             let ident_string = stringify(ident);
                             quote! {
-                                if name == #ident_string {
+                                if __mirror_mirror_name == #ident_string {
                                     return Some(#ident);
                                 }
                             }
@@ -609,7 +626,7 @@ fn expand_enum(
 
         quote! {
             #[allow(unused_variables, unreachable_code)]
-            fn field_mut(&mut self, name: &str) -> Option<&mut dyn Reflect> {
+            fn field_mut(&mut self, __mirror_mirror_name: &str) -> Option<&mut dyn Reflect> {
                 match self {
                     #(#match_arms)*
                     _ => {}
@@ -1038,7 +1055,7 @@ struct NamedField<'a> {
     attrs: InnerAttrs,
 }
 
-impl<'a> NamedField<'a> {
+impl NamedField<'_> {
     #[allow(clippy::wrong_self_convention)]
     fn from_reflect_with(&self) -> Option<&Ident> {
         self.attrs.from_reflect_with.as_ref()
@@ -1051,7 +1068,7 @@ struct UnnamedField<'a> {
     fake_ident: Ident,
 }
 
-impl<'a> UnnamedField<'a> {
+impl UnnamedField<'_> {
     #[allow(clippy::wrong_self_convention)]
     fn from_reflect_with(&self) -> Option<&Ident> {
         self.attrs.from_reflect_with.as_ref()