+++ crates/mirror-mirror-macros/src/lib.rs
@@ -11,7 +11,6 @@
     clippy::needless_borrow,
     clippy::match_wildcard_for_single_variants,
     clippy::if_let_mutex,
-    clippy::mismatched_target_os,
     clippy::await_holding_lock,
     clippy::match_on_vec_items,
     clippy::imprecise_flops,
@@ -57,7 +56,7 @@ mod derive_reflect;
 /// ```
 /// use mirror_mirror::Reflect;
 ///
-/// #[derive(Reflect, Clone, Debug)]
+/// #[derive(Reflect, Clone, Debug, Default)]
 /// struct Foo {
 ///     a: i32,
 ///     b: bool,
@@ -74,7 +73,7 @@ mod derive_reflect;
 /// ```
 /// use mirror_mirror::Reflect;
 ///
-/// #[derive(Reflect, Clone, Debug)]
+/// #[derive(Reflect, Clone, Debug, Default)]
 /// struct Foo(i32, bool, String);
 /// ```
 ///
@@ -86,6 +85,7 @@ mod derive_reflect;
 /// use mirror_mirror::Reflect;
 ///
 /// #[derive(Reflect, Clone, Debug)]
+/// #[reflect(opt_out(Default))]
 /// enum Foo {
 ///     A(i32),
 ///     B { b: bool },
@@ -97,14 +97,14 @@ mod derive_reflect;
 ///
 /// ## `opt_out`
 ///
-/// By default types are required to implement `Clone` and `Debug`. You can opt-out of these
-/// requirements with `#[reflect(opt_out(Clone, Debug))]`
+/// By default types are required to implement `Clone`, `Debug`, and `Default`. You can opt-out of these
+/// requirements with `#[reflect(opt_out(Clone, Debug, Default))]`
 ///
 /// ```
 /// use mirror_mirror::Reflect;
 ///
 /// #[derive(Reflect)]
-/// #[reflect(opt_out(Debug, Clone))]
+/// #[reflect(opt_out(Debug, Clone, Default))]
 /// struct Foo(i32);
 /// ```
 ///
@@ -116,7 +116,7 @@ mod derive_reflect;
 /// ```
 /// use mirror_mirror::{Reflect, FromReflect};
 ///
-/// #[derive(Reflect, Debug, Clone)]
+/// #[derive(Reflect, Debug, Clone, Default)]
 /// #[reflect(opt_out(FromReflect))]
 /// struct Foo(i32);
 ///
@@ -135,16 +135,17 @@ mod derive_reflect;
 /// ```
 /// use mirror_mirror::{Reflect, FromReflect};
 ///
-/// #[derive(Reflect, Debug, Clone)]
+/// #[derive(Reflect, Debug, Clone, Default)]
 /// struct Foo {
 ///     #[reflect(skip)]
 ///     not_reflect: NotReflect,
 /// }
 ///
-/// #[derive(Reflect, Debug, Clone)]
+/// #[derive(Reflect, Debug, Clone, Default)]
 /// struct Bar(#[reflect(skip)] NotReflect);
 ///
 /// #[derive(Reflect, Debug, Clone)]
+/// #[reflect(opt_out(Default))]
 /// enum Baz {
 ///     #[reflect(skip)]
 ///     OnVariant(NotReflect),
@@ -170,7 +171,7 @@ mod derive_reflect;
 /// ```
 /// use mirror_mirror::{Reflect, FromReflect};
 ///
-/// #[derive(Reflect, Debug, Clone)]
+/// #[derive(Reflect, Debug, Clone, Default)]
 /// struct Foo {
 ///     #[reflect(from_reflect_with(n_from_reflect))]
 ///     n: i32,
@@ -194,7 +195,7 @@ mod derive_reflect;
 ///     type_info::{GetMeta, DescribeType},
 /// };
 ///
-/// #[derive(Reflect, Debug, Clone)]
+/// #[derive(Reflect, Debug, Clone, Default)]
 /// #[reflect(meta(
 ///     // a comma separated list of `key = value` pairs.
 ///     //
@@ -236,7 +237,7 @@ mod derive_reflect;
 /// # use mirror_mirror as some_library;
 /// use some_library::Reflect;
 ///
-/// #[derive(Reflect, Debug, Clone)]
+/// #[derive(Reflect, Debug, Clone, Default)]
 /// #[reflect(crate_name(some_library))]
 /// struct Foo {
 ///     n: i32,