+++ crates/mirror-mirror/src/array.rs
@@ -17,6 +17,13 @@ pub trait Array: Reflect {
     fn iter(&self) -> Iter<'_>;
 
     fn iter_mut(&mut self) -> ValueIterMut<'_>;
+
+    /// Swaps two elements in the array.
+    ///
+    /// # Panics
+    ///
+    /// Panics if `a` or `b` are out of bounds.
+    fn swap(&mut self, a: usize, b: usize);
 }
 
 impl fmt::Debug for dyn Array {
@@ -47,10 +54,10 @@ impl<'a> Iterator for Iter<'a> {
     }
 }
 
-impl<'a> ExactSizeIterator for Iter<'a> {
+impl ExactSizeIterator for Iter<'_> {
     fn len(&self) -> usize {
         self.array.len()
     }
 }
 
-impl<'a> FusedIterator for Iter<'a> {}
+impl FusedIterator for Iter<'_> {}