+++ crates/mirror-mirror/src/foreign_impls/boxed.rs
@@ -1,4 +1,3 @@
-use alloc::borrow::Cow;
 use alloc::boxed::Box;
 use core::any::Any;
 use core::fmt;
@@ -12,7 +11,6 @@ use crate::Reflect;
 use crate::ReflectMut;
 use crate::ReflectOwned;
 use crate::ReflectRef;
-use crate::TypeDescriptor;
 use crate::Value;
 
 impl<T> DescribeType for Box<T>
@@ -28,16 +26,16 @@ impl<T> Reflect for Box<T>
 where
     T: Reflect + DescribeType,
 {
-    fn type_descriptor(&self) -> Cow<'static, TypeDescriptor> {
-        <T as DescribeType>::type_descriptor()
-    }
-
     fn as_any(&self) -> &dyn Any {
         <T as Reflect>::as_any(self)
     }
 
     fn as_any_mut(&mut self) -> &mut dyn Any {
         <T as Reflect>::as_any_mut(self)
+    }
+
+    fn into_any(self: Box<Self>) -> Box<dyn Any> {
+        <T as Reflect>::into_any(*self)
     }
 
     fn as_reflect(&self) -> &dyn Reflect {