+++ CHANGELOG.md
@@ -7,7 +7,37 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 # Unreleased
 
-- None.
+- **breaking:** Remove `Reflect::type_descriptor`. Instead capture the type
+  descriptor explicitly with `<T as DescribeType>::type_descriptor()` ([#90])
+- **breaking:** Make `KeyPath::pop` return the popped key ([#110])
+- **breaking:** Remove `Reflect::type_id` ([#109])
+- **breaking:** `no-std` support has been removed ([#119])
+- **fixed:** Iterating over fields and variants in type descriptors is now done
+  in the same order as defined in the code ([#119])
+- **fixed:** Support derive reflect for `HashMap` ([#119])
+- **breaking:** Add `Array::swap` ([#136])
+- **breaking:** Add `List::try_insert` ([#136])
+- **breaking:** `List::push` renamed to `List::try_push` ([#136])
+- **breaking:** `Map::insert` renamed to `Map::try_insert` ([#136])
+- **breaking:** `Map::remove` renamed to `Map::try_remove` ([#136])
+- **breaking:** Update to kollect `0.5`
+- **breaking:** Add reflected sets ([#140])
+- **fixed:** Fix `<[T; N]>::from_reflect` for `Value::List`. It used to fail but now works
+- **fixed:** `Vec::as_array`, `Vec::as_array_mut`, `Vec::into_array` now
+  correctly return `Some(_)`.
+- **fixed:** `Value::as_array`, `Value::as_array_mut`, `Value::into_array` now
+  correctly return `Some(_)`.
+- **breaking:** Require `glam` version from `0.28` to `0.30`
+- **breaking:** Require `glam` version to be >= `0.30.3` when `speedy` feature is enabled
+- **changed:** No longer require `Sized` for `SimpleTypeName::new_from_type`
+- **added:** Support pretty printing trait objects
+
+[#90]: https://github.com/EmbarkStudios/mirror-mirror/pull/90
+[#110]: https://github.com/EmbarkStudios/mirror-mirror/pull/110
+[#109]: https://github.com/EmbarkStudios/mirror-mirror/pull/109
+[#119]: https://github.com/EmbarkStudios/mirror-mirror/pull/119
+[#136]: https://github.com/EmbarkStudios/mirror-mirror/pull/136
+[#140]: https://github.com/EmbarkStudios/mirror-mirror/pull/140
 
 # 0.1.20 (15. April, 2023)
 
@@ -61,6 +91,28 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 # 0.1.11 (20. March, 2023)
 
+- **added:** Implement serialization traits for `ScalarOwned`
+
+# 0.1.14 (11. April, 2023)
+
+- **change:** Make use of deterministic hashes for `NodeId` and `TypeDescriptor` ([#115])
+
+[#115]: https://github.com/EmbarkStudios/mirror-mirror/pull/115
+
+# 0.1.13 (29. March, 2023)
+
+- **fixed:** Require less strict `speedy` version ([#114])
+
+[#114]: https://github.com/EmbarkStudios/mirror-mirror/pull/114
+
+# 0.1.12 (21. March, 2023)
+
+- **change:** Update to syn 2.0 ([#113])
+
+[#113]: https://github.com/EmbarkStudios/mirror-mirror/pull/113
+
+# 0.1.11 (20. March, 2023)
+
 - **added:** Implement `Reflect`, and friends, for `Infallible` ([#111])
 - **change:** Update to syn 2.0 ([#112])
 
@@ -71,6 +123,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - **fixed:** Fully qualify `FromReflect` in generated code ([#107])
 
+[#90]: https://github.com/EmbarkStudios/mirror-mirror/pull/90
 [#107]: https://github.com/EmbarkStudios/mirror-mirror/pull/107
 
 # 0.1.9 (24. February, 2023)